+++ src/python/pants/backend/experimental/go/register.py
@@ -1,17 +1,9 @@
 # Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.backend.go import (
-    build,
-    distribution,
-    import_analysis,
-    module,
-    pkg,
-    sdk,
-    tailor,
-    target_type_rules,
-)
+from pants.backend.go import build, import_analysis, module, pkg, sdk, tailor, target_type_rules
 from pants.backend.go import target_types as go_target_types
+from pants.backend.go.subsystems import golang
 from pants.backend.go.target_types import GoBinary, GoExternalPackageTarget, GoModule, GoPackage
 
 
@@ -22,7 +14,7 @@ def target_types():
 def rules():
     return [
         *build.rules(),
-        *distribution.rules(),
+        *golang.rules(),
         *go_target_types.rules(),
         *import_analysis.rules(),
         *module.rules(),
