+++ src/python/pants/backend/go/import_analysis.py
@@ -8,7 +8,7 @@ from typing import Dict
 
 import ijson
 
-from pants.backend.go.distribution import GoLangDistribution
+from pants.backend.go.subsystems.golang import GoLangDistribution
 from pants.core.util_rules.external_tool import DownloadedExternalTool, ExternalToolRequest
 from pants.engine.fs import CreateDigest, Digest, FileContent, MergeDigests
 from pants.engine.internals.selectors import Get
